+++ NEWS.md
@@ -12,6 +12,7 @@ instead of platform
 * add record argument to edit* functions to preserve the series
 of actions from an editing session.  If `record = TRUE` then a `"recorder"` attribute will be added to the returned object for
 full reproducibility.
+* add internal playback for recorded session for future use
 * add viewer argument to select and edit functions to allow
 user the flexibility to adjust the viewer experience.  Default
 will be paneViewer() in an attempt to keep the workflow